fix lipi lekhika typing
