+++ app/src/components/pages/map_edit/EditPathLabel.svelte
