+++ app/src/components/pages/map_edit/ProjectMapEditor.svelte
@@ -1,15 +1,5 @@
 <script lang="ts">
-  import './map_edit_tree.css';
-  import {
-    Map as MapIcon,
-    TreePine,
-    PenLine,
-    ArrowUpDown,
-    Ban,
-    GripVertical,
-    FolderRoot
-  } from '@lucide/svelte';
-  import { Tree } from '@keenmate/svelte-treeview';
+  import { Map as MapIcon, ArrowUpDown, FolderRoot } from '@lucide/svelte';
   import type { Tree as TreeComponent, LTreeNode, DropPosition } from '@keenmate/svelte-treeview';
   import { browser } from '$app/environment';
   import { beforeNavigate, goto } from '$app/navigation';
@@ -25,25 +15,16 @@
   import { map_edit_dirty } from '~/state/map_edit_dirty.svelte';
   import type { recursive_list_type } from '~/state/data_types';
   import { Button } from '$lib/components/ui/button';
-  import { Input } from '$lib/components/ui/input';
-  import { Label } from '$lib/components/ui/label';
   import { Badge } from '$lib/components/ui/badge';
   import * as Card from '$lib/components/ui/card';
-  import * as ScrollArea from '$lib/components/ui/scroll-area';
   import * as Breadcrumb from '$lib/components/ui/breadcrumb';
   import * as AlertDialog from '$lib/components/ui/alert-dialog';
-  import { Separator } from '$lib/components/ui/separator';
-  import { Switch } from '$lib/components/ui/switch';
   import { toast } from 'svelte-sonner';
-  import {
-    clearTypingContextOnKeyDown,
-    createTypingContext,
-    handleTypingBeforeInputEvent
-  } from 'lipilekhika/typing';
   import {
     type MapNodeWithClientId,
     type MapPath,
     type BaselineNodeSnapshot,
+    type MapTreeItem,
     clone_map_with_client_ids,
     strip_client_ids,
     parse_path_query,
@@ -67,9 +48,10 @@
     expand_tree_expanded_paths,
     collapse_tree_expanded_paths
   } from './map_edit_lib';
-  import MapEditPathLabel from './MapEditPathLabel.svelte';
-  import * as Tooltip from '$lib/components/ui/tooltip';
   import type { PathSwapEdit } from '~/server/map_path_swap';
+  import TreeEditPanel from './TreeEditPanel.svelte';
+  import NodeEditor from './NodeEditor.svelte';
+  import ChangesPanel from './ChangesPanel.svelte';
 
   let {
     project_id,
@@ -95,41 +77,10 @@
   let count_input_invalid = $state(false);
   let treeRef = $state<TreeComponent<MapTreeItem> | undefined>(undefined);
   let saving_order = $state(false);
-  let typing_enabled = $state(true);
   let last_tree_click = $state<{ path: string; at: number } | null>(null);
   let expandedTreePaths = $state<Set<string>>(default_tree_expanded_paths());
 
   const TREE_DBLCLICK_MS = 400;
-
-  const typing_ctx = $derived(
-    createTypingContext('Devanagari', {
-      includeInherentVowel: true
-    })
-  );
-
-  $effect(() => {
-    typing_ctx.ready;
-  });
-
-  type MapTreeItem = {
-    path: string;
-    id: string;
-    name_dev: string;
-    level: number;
-    parentPath: string;
-    sortOrder: number;
-    visibleIndex: number;
-    isRoot: boolean;
-    isLeaf: boolean;
-    nodeType: 'list' | 'shloka';
-    childCount: number;
-    edited: boolean;
-    moved: boolean;
-    draggable: boolean;
-    isSelected: boolean;
-    isExpanded: boolean;
-    allowedDropPositions: ('above' | 'below')[];
-  };
 
   const diffState = $derived.by(() => {
     if (!workingMap) {
@@ -358,15 +309,6 @@
     e.stopPropagation();
     e.preventDefault();
     promote_change_root(subtreePath);
-  }
-
-  function toggle_typing_from_keyboard(e: KeyboardEvent) {
-    if (!e.altKey) return false;
-    const key = e.key.toLowerCase();
-    if (key !== 'x' && key !== 'c') return false;
-    e.preventDefault();
-    typing_enabled = !typing_enabled;
-    return true;
   }
 
   function update_name_dev(value: string) {
@@ -651,7 +593,20 @@
         </Breadcrumb.List>
       </Breadcrumb.Root>
 
-      {#if order_edit_mode}
+      {#if !order_edit_mode}
+        <div class="flex flex-wrap items-center gap-3 rounded-md bg-muted/40 px-3 py-1.5">
+          <p class="text-xs text-muted-foreground">
+            Hover a branch with children and use
+            <FolderRoot class="inline size-3 align-text-bottom text-primary" />
+            to set tree root. Double-click a node to expand or collapse.
+          </p>
+          {#if base_path_active}
+            <Button variant="ghost" size="sm" class="h-6 text-xs" onclick={() => set_base_path([])}>
+              ← Back to full tree
+            </Button>
+          {/if}
+        </div>
+      {:else}
         <div class="rounded-md bg-amber-500/8 px-3 py-1.5 dark:bg-amber-400/8">
           <p class="text-xs text-amber-800 dark:text-amber-300">
             Order edit mode — drag siblings to reorder. Field edits are locked.
@@ -661,367 +616,40 @@
     </Card.Header>
   </Card.Root>
 
-  {#if !order_edit_mode}
-    <div
-      class="flex flex-wrap items-center justify-between gap-3 rounded-lg border border-border/60 bg-muted/20 px-4 py-2.5"
-    >
-      <div class="flex items-center gap-2">
-        <div
-          class="flex size-5 shrink-0 items-center justify-center rounded text-primary opacity-70"
-        >
-          <PenLine class="size-3.5" />
-        </div>
-        <p class="text-xs text-muted-foreground">
-          Lipi-Lekhika typing for <span class="font-medium text-foreground">Name (Devanagari)</span>
-          — toggle with
-          <kbd
-            class="rounded border border-border bg-background px-1.5 font-mono text-[10px] shadow-sm"
-            >Alt+X</kbd
-          >
-          or
-          <kbd
-            class="rounded border border-border bg-background px-1.5 font-mono text-[10px] shadow-sm"
-            >Alt+C</kbd
-          >
-        </p>
-      </div>
-      <div class="flex items-center gap-2">
-        <Label
-          for="map-edit-typing"
-          class="cursor-pointer text-xs font-medium text-muted-foreground select-none"
-        >
-          Typing
-        </Label>
-        <Switch
-          id="map-edit-typing"
-          checked={typing_enabled}
-          onCheckedChange={(v) => (typing_enabled = v)}
-          title="Devanagari transliteration typing (Alt+X / Alt+C)"
-        />
-      </div>
-    </div>
-  {/if}
-
   <div class="grid gap-4 lg:grid-cols-[minmax(0,1fr)_minmax(0,1.15fr)]">
-    <Card.Root class="flex min-h-[420px] flex-col overflow-hidden lg:min-h-[min(72vh,640px)]">
-      <div class="border-b border-border/60 bg-muted/20 px-5 py-3">
-        <div class="flex items-center gap-2">
-          <TreePine class="size-4 shrink-0 text-muted-foreground" />
-          <span class="text-sm font-semibold">Tree</span>
-        </div>
-        <p class="mt-0.5 text-xs text-muted-foreground">
-          {#if order_edit_mode}
-            Drag rows to reorder siblings — subtree under {base_path_resolved}
-          {:else}
-            Subtree under {base_path_resolved}
-          {/if}
-        </p>
-      </div>
-      <Card.Content class="min-h-0 flex-1 px-3 pt-3 pb-4">
-        <ScrollArea.Root class="h-[min(52vh,420px)] lg:h-[min(60vh,560px)]">
-          {#if workingMap && treeData.length > 0}
-            <div class="map-edit-tree px-1">
-              <Tree
-                bind:this={treeRef}
-                data={treeData}
-                idMember="id"
-                pathMember="path"
-                parentPathMember="parentPath"
-                levelMember="level"
-                displayValueMember="name_dev"
-                orderMember="sortOrder"
-                isSelectedMember="isSelected"
-                isExpandedMember="isExpanded"
-                shouldToggleOnNodeClick={false}
-                dragDropMode={order_edit_mode ? 'self' : 'none'}
-                allowedDropPositionsMember="allowedDropPositions"
-                getAllowedDropPositionsCallback={order_edit_mode
-                  ? get_allowed_drop_positions
-                  : undefined}
-                beforeDropCallback={order_edit_mode ? before_drop : undefined}
-                onNodeClicked={on_tree_node_clicked}
-              >
-                {#snippet nodeTemplate(node: LTreeNode<MapTreeItem>)}
-                  {@const row = node.data!}
-                  {@const nodeKind = order_edit_mode
-                    ? 'neutral'
-                    : row.nodeType === 'shloka'
-                      ? 'leaf'
-                      : row.childCount === 0
-                        ? 'empty-list'
-                        : 'list'}
-                  <div
-                    class="map-edit-tree-row group flex w-full items-center gap-2 py-1 pr-2"
-                    class:map-edit-row-selected={row.isSelected}
-                    data-node-kind={nodeKind}
-                  >
-                    {#if order_edit_mode}
-                      {#if row.draggable}
-                        <span
-                          class="text-muted-foreground/60 transition-colors hover:text-muted-foreground"
-                          title="Drag to reorder"
-                        >
-                          <GripVertical class="size-3.5" />
-                        </span>
-                        <span class="w-5 shrink-0 text-[11px] text-muted-foreground tabular-nums">
-                          {row.visibleIndex}.
-                        </span>
-                      {:else if !row.isRoot}
-                        <span class="w-8 shrink-0" aria-hidden="true"></span>
-                      {/if}
-                    {:else if !row.isRoot}
-                      {#if row.nodeType === 'list' && row.childCount === 0}
-                        <span
-                          class="shrink-0 text-red-400 dark:text-red-500"
-                          title="Empty list — no children"
-                        >
-                          <Ban class="size-3" />
-                        </span>
-                      {/if}
-                      <span class="w-5 shrink-0 text-[11px] text-muted-foreground/60 tabular-nums">
-                        {row.visibleIndex}.
-                      </span>
-                    {/if}
-                    <span
-                      class="min-w-0 flex-1 truncate text-sm {row.nodeType === 'list' &&
-                      row.childCount === 0 &&
-                      !row.isRoot &&
-                      !order_edit_mode
-                        ? 'opacity-60'
-                        : ''}">{row.name_dev}</span
-                    >
-                    <div class="flex shrink-0 items-center gap-1">
-                      {#if !order_edit_mode &&
-                        row.nodeType === 'list' &&
-                        row.childCount > 0 &&
-                        !row.isRoot}
-                        <button
-                          type="button"
-                          class="map-edit-set-root-btn invisible shrink-0 rounded p-0.5 text-muted-foreground transition-colors hover:bg-muted hover:text-primary"
-                          title="Set as tree root"
-                          aria-label="Set as tree root"
-                          onclick={(e) => on_set_tree_root_click(e, row.path)}
-                        >
-                          <FolderRoot class="size-3.5" />
-                        </button>
-                      {/if}
-                      {#if row.isRoot}
-                        <Badge variant="outline" class="border-primary/30 text-[10px] text-primary"
-                          >root</Badge
-                        >
-                      {/if}
-                      {#if !order_edit_mode && row.edited}
-                        <Badge
-                          variant="secondary"
-                          class="bg-amber-100 text-[10px] text-amber-700 dark:bg-amber-900/30 dark:text-amber-400"
-                          >edited</Badge
-                        >
-                      {/if}
-                      {#if row.moved}
-                        <Badge class="text-[10px]">moved</Badge>
-                      {/if}
-                    </div>
-                  </div>
-                {/snippet}
-              </Tree>
-            </div>
-          {:else}
-            <p class="p-4 text-sm text-muted-foreground">Loading map…</p>
-          {/if}
-        </ScrollArea.Root>
-      </Card.Content>
-    </Card.Root>
-
-    <Card.Root class="flex min-h-[420px] flex-col overflow-hidden lg:min-h-[min(72vh,640px)]">
-      <div class="border-b border-border/60 bg-muted/20 px-5 py-3">
-        <div class="flex items-center gap-2">
-          <PenLine class="size-4 shrink-0 text-muted-foreground" />
-          <span class="text-sm font-semibold">Node editor</span>
-        </div>
-        <p class="mt-0.5 text-xs text-muted-foreground">
-          {#if selectedNode}
-            <span class="text-foreground">{selected_path_resolved}</span>
-          {:else}
-            Select a node in the tree
-          {/if}
-        </p>
-      </div>
-      <Card.Content class="space-y-4 pt-4">
-        {#if order_edit_mode}
-          <div class="rounded-lg border border-border/50 bg-muted/30 px-4 py-3">
-            <p class="text-sm text-muted-foreground">
-              {#if selectedNode}
-                Viewing <span class="font-medium text-foreground">{selected_path_resolved}</span>.
-                Reorder in the tree; names and list labels cannot be edited in this mode.
-              {:else}
-                Select a node to inspect it. Reorder siblings using drag handles in the tree.
-              {/if}
-            </p>
-          </div>
-        {:else if selectedNode}
-          <div class="space-y-1.5">
-            <Label
-              for="name_dev"
-              class="text-xs font-medium tracking-wide text-muted-foreground uppercase"
-              >Name (Devanagari)</Label
-            >
-            <Input
-              id="name_dev"
-              value={selectedNode.name_dev}
-              disabled={selected_is_root}
-              onbeforeinput={(e) =>
-                handleTypingBeforeInputEvent(
-                  typing_ctx,
-                  e,
-                  (newValue) => update_name_dev(newValue),
-                  typing_enabled
-                )}
-              onblur={() => typing_ctx.clearContext()}
-              onkeydown={(e) => {
-                if (toggle_typing_from_keyboard(e)) return;
-                clearTypingContextOnKeyDown(e, typing_ctx);
-              }}
-            />
-            {#if selected_is_root}
-              <p class="text-xs text-muted-foreground">
-                Root display name (Devanagari) is not edited here.
-              </p>
-            {/if}
-          </div>
-
-          {#if selectedNode.info.type === 'list'}
-            <Separator />
-            <div class="space-y-1.5">
-              <Label
-                for="list_name"
-                class="text-xs font-medium tracking-wide text-muted-foreground uppercase"
-                >List type label</Label
-              >
-              <Input
-                id="list_name"
-                value={selectedNode.info.list_name}
-                oninput={(e) => update_list_name(e.currentTarget.value)}
-              />
-              {#if selected_is_root}
-                <p class="text-xs text-muted-foreground">
-                  Top-level list type label (e.g. Veda, Kanda). Editable at the project root only.
-                </p>
-              {/if}
-            </div>
-            <div class="space-y-1.5">
-              <Label
-                for="list_count_expected"
-                class="text-xs font-medium tracking-wide text-muted-foreground uppercase"
-                >Expected list count <span class="font-normal normal-case">(optional)</span></Label
-              >
-              <Input
-                id="list_count_expected"
-                inputmode="numeric"
-                bind:value={list_count_draft}
-                oninput={(e) => update_list_count_expected(e.currentTarget.value)}
-                aria-invalid={count_input_invalid}
-              />
-              {#if count_input_invalid}
-                <p class="text-xs text-destructive">Enter a whole number ≥ 0, or leave empty.</p>
-              {/if}
-            </div>
-          {:else}
-            <Separator />
-            <div class="rounded-lg border border-border/50 bg-muted/20 p-3 text-sm">
-              <p class="mb-2 text-xs font-medium tracking-wide text-muted-foreground uppercase">
-                Shloka metadata
-              </p>
-              <dl class="grid grid-cols-2 gap-x-3 gap-y-1.5 text-sm">
-                <dt class="text-muted-foreground">Shloka count</dt>
-                <dd class="font-medium tabular-nums">{selectedNode.info.shloka_count}</dd>
-                <dt class="text-muted-foreground">Total lines</dt>
-                <dd class="font-medium tabular-nums">{selectedNode.info.total}</dd>
-                {#if selectedNode.info.shloka_count_expected != null}
-                  <dt class="text-muted-foreground">Expected</dt>
-                  <dd class="font-medium tabular-nums">
-                    {selectedNode.info.shloka_count_expected}
-                  </dd>
-                {/if}
-              </dl>
-              <p class="mt-2 text-xs text-muted-foreground">
-                Derived fields update when texts change; not editable here.
-              </p>
-            </div>
-          {/if}
-        {:else}
-          <div class="flex h-full flex-col items-center justify-center py-8 text-center">
-            <div class="mb-3 flex size-10 items-center justify-center rounded-full bg-muted/60">
-              <PenLine class="size-4 text-muted-foreground" />
-            </div>
-            <p class="text-sm text-muted-foreground">
-              Choose a node from the tree to edit its fields.
-            </p>
-          </div>
-        {/if}
-      </Card.Content>
-    </Card.Root>
+    <TreeEditPanel
+      {workingMap}
+      {treeData}
+      {order_edit_mode}
+      {base_path_resolved}
+      bind:treeRef
+      onNodeClicked={on_tree_node_clicked}
+      beforeDrop={before_drop}
+      getAllowedDropPositions={get_allowed_drop_positions}
+      onSetTreeRoot={on_set_tree_root_click}
+    />
+    <NodeEditor
+      {order_edit_mode}
+      {selectedNode}
+      {selected_path_resolved}
+      {selected_is_root}
+      bind:list_count_draft
+      {count_input_invalid}
+      onNameDevChange={update_name_dev}
+      onListNameChange={update_list_name}
+      onListCountChange={update_list_count_expected}
+    />
   </div>
 
-  <Card.Root class="overflow-hidden">
-    <div class="flex items-center justify-between border-b border-border/60 bg-muted/20 px-5 py-3">
-      <div>
-        <span class="text-sm font-semibold">
-          {order_edit_mode ? 'Order changes' : 'Changes'}
-        </span>
-        <span class="ml-2 text-xs text-muted-foreground">
-          {#if order_edit_mode}
-            {#if order_dirty}
-              {pending_swaps.length} swap{pending_swaps.length === 1 ? '' : 's'} pending
-            {:else}
-              no changes yet
-            {/if}
-          {:else if metadata_dirty}
-            {diffState.rows.length} update{diffState.rows.length === 1 ? '' : 's'}
-          {:else}
-            no unsaved changes
-          {/if}
-        </span>
-      </div>
-      {#if order_edit_mode ? order_dirty : metadata_dirty}
-        <div class="size-2 animate-pulse rounded-full bg-primary"></div>
-      {/if}
-    </div>
-    <Card.Content class="pt-4 pb-4">
-      {#if order_edit_mode}
-        {#if active_diff_state.rows.length === 0 && pending_swaps.length === 0}
-          <p class="text-sm text-muted-foreground">Drag siblings in the tree to reorder lists.</p>
-        {:else if workingMap}
-          <Tooltip.Provider>
-            <ul class="grid gap-2 sm:grid-cols-2 lg:grid-cols-3">
-              {#each active_diff_state.rows as row (row.kind + row.clientId + row.summary)}
-                <li
-                  class="space-y-1 rounded-md border border-primary/20 bg-primary/5 px-2.5 py-2 text-xs leading-snug text-foreground"
-                >
-                  <MapEditPathLabel map={workingMap} path={row.path} variant="short" />
-                  <p>{row.summary}</p>
-                </li>
-              {/each}
-            </ul>
-          </Tooltip.Provider>
-        {/if}
-      {:else if active_diff_state.rows.length === 0}
-        <p class="text-sm text-muted-foreground">Edits will appear here.</p>
-      {:else if workingMap}
-        <Tooltip.Provider>
-          <ul class="grid gap-2 sm:grid-cols-2 lg:grid-cols-3">
-            {#each active_diff_state.rows as row (row.kind + row.clientId + row.summary)}
-              <li
-                class="space-y-1 rounded-md border border-primary/20 bg-primary/5 px-2.5 py-2 text-xs leading-snug text-foreground"
-              >
-                <MapEditPathLabel map={workingMap} path={row.path} variant="short" />
-                <p>{row.summary}</p>
-              </li>
-            {/each}
-          </ul>
-        </Tooltip.Provider>
-      {/if}
-    </Card.Content>
-  </Card.Root>
+  <ChangesPanel
+    {order_edit_mode}
+    {order_dirty}
+    {metadata_dirty}
+    {active_diff_state}
+    {pending_swaps}
+    {diffState}
+    {workingMap}
+  />
 </div>
 
 <AlertDialog.Root bind:open={save_dialog_open}>
